db_search_agent/adk-agent/requirements.txt
# Google ADK and core dependencies
google-adk>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# PostgreSQL adapter for direct database access
psycopg2-binary>=2.9.9

# Environment variable management
python-dotenv>=1.0.0

# Optional: AgentOps for observability (uncomment if needed)
# agentops>=1.0.0

